Mark addr connected when connecting to self
Prior to this patch, notbit would repeatedly attempt to connect to
itself if it's own addr was in the address store since it believes it is
not connected. This causes quite a bit of log spam. A rememdy to this
problem is to mark the address as connected. This causes notbit to ignore
the address when searching for new peers with which to establish a connection.